new containers; reorg zone -> buddy code; buddy system works with page struct; rename some memory allocators; smp has a new way to startup AP; bug fix; get core/threads count from cpuid;

+++ naos/includes/kernel/util/hash.hpp
@@ -3,7 +3,7 @@
 namespace util
 {
 u32 djb_hash(const char *str);
-u64 murmur_hash2_64(const void *key, int len, u64 seed);
-void murmur_hash3_128(const void *key, const int len, const uint32_t seed, void *out);
+u64 murmur_hash2_64(const void *key, u64 len, u64 seed);
+void murmur_hash3_128(const void *key, const u64 len, const uint32_t seed, void *out);
 
 } // namespace util
